+++ CHANGELOG.md
@@ -8,8 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [2.32.0] - 2022-06-03
 
 ### Added
-- Add support for hosting meetings in US GovCloud regions.
-- Add support for starting live transcription in US GovCloud regions.
+
 ### Removed
 
 ### Changed